refactor(repository): :recycle: Extracted settings specific functions into own file
Extract settings related stuff from userRepository and move them to new settingsRepository file.
