+++ src/repository/userRepository.ts
@@ -1,4 +1,4 @@
-import { AuthUser, SettingsData } from '@/types';
+import { AuthUser } from '@/types';
 import { prisma } from './prisma';
 import * as Sentry from '@sentry/nextjs';
 
@@ -27,71 +27,6 @@ export async function getUser(email: string) {
         where: { email },
       });
       return user;
-    },
-  );
-}
-export async function getSettings(userId: number) {
-  return await Sentry.startSpan(
-    { name: 'getSettings', op: 'db.sql.prisma' },
-    async () => {
-      const s = await prisma.settings.findUnique({
-        where: {
-          user_id: userId,
-        },
-      });
-      return s;
-    },
-  );
-}
-export async function insertSettings(
-  userId: number,
-  beginDate: Date,
-  initialBalanceHours: number,
-  initialBalanceMins: number,
-) {
-  return await Sentry.startSpan(
-    { name: 'insertSettings', op: 'db.sql.prisma' },
-    async () => {
-      const s = await prisma.settings.upsert({
-        where: {
-          user_id: userId,
-        },
-        create: {
-          user_id: userId,
-          begin_date: beginDate,
-          initial_balance_hours: initialBalanceHours,
-          initial_balance_mins: initialBalanceMins,
-        },
-        update: {},
-      });
-      return s;
-    },
-  );
-}
-export async function upsertSettings(
-  userId: number,
-  { beginDate, initialBalanceHours, initialBalanceMins }: SettingsData,
-) {
-  return await Sentry.startSpan(
-    { name: 'upsertSettings', op: 'db.sql.prisma' },
-    async () => {
-      const s = await prisma.settings.upsert({
-        where: {
-          user_id: userId,
-        },
-        create: {
-          user_id: userId,
-          begin_date: beginDate,
-          initial_balance_hours: initialBalanceHours,
-          initial_balance_mins: initialBalanceMins,
-        },
-        update: {
-          begin_date: beginDate,
-          initial_balance_hours: initialBalanceHours,
-          initial_balance_mins: initialBalanceMins,
-        },
-      });
-      return s;
     },
   );
 }
